:package: Remove nedpals/supabase-go dependency
